requirements.txt
torch
transformers
gradio
Pillow
fpdf